+++ packages/adb/lib/src/adb.dart
@@ -91,6 +91,7 @@ Future<void> instrument({
   String? device,
   void Function(String)? onStdout,
   void Function(String)? onStderr,
+  Map<String, String> arguments = const {},
 }) async {
   final process = await Process.start(
     'adb',
@@ -103,6 +104,11 @@ Future<void> instrument({
       'am',
       'instrument',
       '-w',
+      for (final arg in arguments.entries) ...[
+        '-e',
+        arg.key,
+        arg.value,
+      ],
       '$packageName/$intentClass',
     ],
     runInShell: true,