+++ packages/maestro_cli/lib/src/features/drive/constants.dart
@@ -1,0 +1,3 @@
+const envHostKey = 'MAESTRO_HOST';
+const envPortKey = 'MAESTRO_PORT';
+const envVerboseKey = 'MAESTRO_VERBOSE';