+++ packages/maestro_test/example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 
 void main() {
   runApp(const MyApp());
@@ -29,12 +30,38 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
+  var _counter = 0;
+  final _notificationsPlugin = FlutterLocalNotificationsPlugin();
+
+  @override
+  void initState() {
+    super.initState();
+
+    _notificationsPlugin.initialize(
+      const InitializationSettings(
+        android: AndroidInitializationSettings('@mipmap/ic_launcher'),
+      ),
+    );
+  }
 
   void _incrementCounter() {
     setState(() {
       _counter++;
     });
+  }
+
+  void _showNotification() {
+    _notificationsPlugin.show(
+      1,
+      'Maestro example',
+      'Hello there!',
+      const NotificationDetails(
+        android: AndroidNotificationDetails(
+          'main',
+          'Default notification channel',
+        ),
+      ),
+    );
   }
 
   @override
@@ -55,6 +82,14 @@ class _MyHomePageState extends State<MyHomePage> {
               key: const ValueKey('counterText'),
               style: Theme.of(context).textTheme.headline4,
             ),
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headline4,
+            ),
+            TextButton(
+              onPressed: _showNotification,
+              child: const Text('Show notification'),
+            )
           ],
         ),
       ),