+++ packages/maestro_test/example/test_driver/integration_test.dart
@@ -3,14 +3,16 @@ import 'package:maestro_test/maestro_test.dart';
 
 // Runs on our machine. Knows nothing about the app being tested.
 Future<void> main() async {
-  print('Waiting for automator server');
-  Automator.init(verbose: true);
-  while (!await Automator.instance.isRunning()) {}
-  print('Automator server is running, starting test drive');
+  final maestro = Maestro.forDriver();
+  while (!await maestro.isRunning()) {
+    print('Waiting for maestro automation server...');
+    await Future<void>.delayed(const Duration(seconds: 1));
+  }
+  print('Maestro automation server is running, starting test drive');
   try {
     await integrationDriver();
   } finally {
-    print('Stopping automator server');
-    await Automator.instance.stop();
+    print('Stopping Maestro automation server');
+    await maestro.stop();
   }
 }