+++ packages/maestro_test/lib/src/maestro.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:http/http.dart' as http;
 import 'package:logging/logging.dart' as logging;
@@ -9,17 +10,41 @@ import 'package:maestro_test/src/extensions.dart';
 ///
 /// Communicates over HTTP with the Maestro server app running on the target
 /// device.
-class Automator {
-  Automator._();
+class Maestro {
+  /// Creates a new [Maestro] instance for use in the driver file (on test
+  /// host).
+  Maestro.forDriver()
+      : host = Platform.environment['MAESTRO_HOST']!,
+        port = Platform.environment['MAESTRO_PORT']!,
+        verbose = Platform.environment['MAESTRO_VERBOSE'] == 'true' {
+    print(
+      'Creating Maestro driver instance. Host: $host, port: $port, verbose: $verbose',
+    );
+    _setUpLogger();
+  }
 
-  static Automator? _instance;
+  /// Creates a new [Maestro] instance for use in testing environment (on target
+  /// device).
+  Maestro.forTest()
+      : host = const String.fromEnvironment('MAESTRO_HOST'),
+        port = const String.fromEnvironment('MAESTRO_PORT'),
+        verbose = const String.fromEnvironment('MAESTRO_VERBOSE') == 'true' {
+    print(
+      'Creating Maestro test instance. Host: $host, port: $port, verbose: $verbose',
+    );
+    _setUpLogger();
+  }
 
-  // ignore: prefer_constructors_over_static_methods
-  static Automator get instance => _instance ??= Automator._();
+  final String host;
+  final String port;
+  final bool verbose;
 
-  static void init({int port = 8081, bool verbose = false}) {
-    instance._port = port;
+  final _client = http.Client();
+  final _logger = logging.Logger('Maestro');
 
+  String get _baseUri => 'http://$host:$port';
+
+  void _setUpLogger() {
     logging.Logger.root.onRecord.listen((event) {
       // ignore: avoid_print
       print('${event.loggerName}: ${event.message}');
@@ -27,17 +52,11 @@ class Automator {
 
     logging.hierarchicalLoggingEnabled = true;
     if (verbose) {
-      instance._logger.level = logging.Level.ALL;
+      _logger.level = logging.Level.ALL;
     } else {
-      instance._logger.level = logging.Level.INFO;
+      _logger.level = logging.Level.INFO;
     }
   }
-
-  final _client = http.Client();
-  final _logger = logging.Logger('Automator');
-
-  late final int _port;
-  String get _baseUri => 'http://localhost:$_port';
 
   Future<http.Response> _wrapPost(
     String action, [
@@ -66,7 +85,7 @@ class Automator {
     try {
       final res = await _client.get(Uri.parse('$_baseUri/healthCheck'));
       _logger.info(
-        'status code: ${res.statusCode}, response body:\n ${res.body}',
+        'status code: ${res.statusCode}, response body: ${res.body}',
       );
       return res.successful;
     } catch (err, st) {
@@ -142,9 +161,16 @@ class Automator {
   /// Disables bluetooth
   Future<void> disableBluetooth() => _wrapPost('disableBluetooth');
 
-  /// Taps at the [index]-th visible button.
+  /// Taps on the [index]-th visible button.
   Future<void> tap(int index) {
     return _wrapPost('tap', <String, dynamic>{'index': index});
+  }
+
+  /// Taps on the [index]-th visible notification.
+  ///
+  /// You must call [openNotifications] first.
+  Future<void> tapOnNotification(int index) {
+    return _wrapPost('tapOnNotification', <String, dynamic>{'index': index});
   }
 
   /// Enters text to the [index]-th visible text field.