+++ packages/adb/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4
+
+- Fix wrong order of arguments for `instrument()`
+
+## 0.0.3
+
+- Make it possible to pass arguments to `instrument()` method
+
 ## 0.0.2
 
 - Fix bugs that arose during migration.